pyproject.toml
[project]
name = "t2v-eval"
version = "0.1.0"
description = "Unified evaluation pipeline for text-to-video generation with official metric implementations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.11"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["text-to-video", "evaluation", "vbench", "clip", "video-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "numpy<2",
    "pandas>=1.3.0",
    "PyYAML>=6.0",
    "tqdm>=4.62.0",
    "Pillow>=8.0.0",
    # HuggingFace
    "datasets>=2.14.0",
    "huggingface-hub>=0.16.0",
    # Video processing (CPU-friendly)
    "decord>=0.6.0",
    "opencv-python-headless>=4.5.0",
    # Utilities
    "gdown>=5.2.1",
    "torch==2.4.*",
    "torchvision==0.19.*",
    "timm<=1.0.12",
    "scenedetect>=0.6.7.1",
    "moviepy>=2.2.1",
    "scipy>=1.15.3",
    "easydict>=1.13",
    "av<14",
    "dreamsim>=0.2.1",
    "omegaconf>=2.3.0",
    "openai-clip>=1.0.1",
    "transformers==4.49.0",
    "setuptools<81",
    "pyiqa>=0.1.10",
    "fairscale>=0.4.4",
    "fvcore>=0.1.5.post20221221",
    "lvis>=0.5.3",
    "wheel>=0.46.3",
    "ninja>=1.13.0",
    "hatchling>=1.28.0",
    "boto3>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

vbench-full = [
    # Heavy deps for VBench 16D object-related dimensions
    "detectron2",
]

vbench = [
    # Minimal deps for temporal_flickering + motion_smoothness (CPU)
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "omegaconf>=2.3.0",
    "imageio>=2.31.1",
]

niqe = [
    # NIQE metrics (optional)
    "pyiqa>=0.1.7",
    "scikit-video>=1.1.11",
    "torch>=2.0.0",
]

t2v-metrics-core = [
    # Minimal deps to import t2v_metrics + run CLIPScore on CPU
    "torch>=2.0.0",
    "open-clip-torch>=2.20.0",
    "transformers==4.49.0",
    "scikit-learn>=1.0.0",
]

all = [
    "t2v-eval[dev,vbench,niqe,t2v-metrics-core]",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/t2v-eval"
Repository = "https://github.com/YOUR_USERNAME/t2v-eval"
Documentation = "https://github.com/YOUR_USERNAME/t2v-eval#readme"

[project.scripts]
t2v-eval = "scripts.run_eval_core:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
[tool.uv.extra-build-dependencies]
detectron2 = ["torch"]
t2v-eval = ["hatchling"]

[tool.uv.sources]
detectron2 = { git = "https://github.com/facebookresearch/detectron2.git" }

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "black>=22.0.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
